requirements.txt.txt
python-telegram-bot==20.3
openai==1.14.3
flask==2.2.5
python-dotenv==1.0.0
